chore: move TracksSidebar & TracksModal
